Capture exception backtraces in Rails3, even if handled by Rails itself
ActionDispatch in Rails 3+ captures exceptions a bit too late for
us to handle them properly within our LogSubscriber alone. And
it seems, someone found the need to not just add the original exception
to the payload in the first place...

As such, we introduce a new exception_app (available since Rails
3.2.0) to capture the exception when it is actually handled by
ActionDispatch and add the backtrace to the Rackstash fields. We then
pass the controll to the original exception_app.