Explicitly set the path in buffer fields for Rails3+
If an exception occured during a request (either from the app itself,
the routing, or anything else) the exception is bubbled up until it
is handled by the ActionDispatch::ShowExceptions middleware.

This middleware generates the actual error page shown to the user.
It also sets the requests's path to "/#{status}" which thus
overrides the original path when we read it in
Rackstash::LogMiddleware#set_fields later.