Collect common request fields before running the app
As Rack::Request just relies on environment variables for most stuff
and the AS::LogSubsubsriber is run asynchronously, we need to collect
the correct data (esp. the path) before the app is able to be "smart"
to change the request fields around.

Downside of this approach is, that we are now unable to reflect
legitimate changes of the environment by later middleware apps.
However, these can explicitly set the Rackstash buffer field
explicitly themselves.